Language.Haskell.GHC.ExactPrint.Print

Description

Documentation

type Anns = Map AnnKey Annotation

exactPrintWithAnns :: Annotate ast => Located ast -> Anns -> String

Print an AST with a map of potential modified Anns. The usual way to generate such a map is by calling relativiseApiAnns.

exactPrint :: Annotate ast => Located ast -> ApiAnns -> String

Print an AST exactly as specified by the annotations on the nodes in the tree. The output of this function should exactly match the source file.