Language.Haskell.GHC.ExactPrint.Types

Documentation

data Comment

A Haskell comment.

Constructors

Comment Span String

data DComment

Delta version of the comment.

Constructors

DComment (DeltaPos, DeltaPos) String

type Pos = (Int, Int)

type Span = (Pos, Pos)

type PosToken = (Located Token, String)

newtype DeltaPos

Constructors

DP (Int, Int)

newtype LayoutStartCol

Marks the start column of a layout block.

Constructors

LayoutStartCol

Fields

getLayoutStartCol :: Int

newtype ColDelta

Marks the distance from the start of the layout block to the element.

Constructors

ColDelta

Fields

getColDelta :: Int

data Annotation

Constructors

Ann

Fields

annEntryDelta :: !DeltaPos

Offset used to get to the start of the SrcSpan.

annDelta :: !ColDelta

Offset from the start of the current layout block. This is used when moving onto new lines when layout rules must be obeyed.

annsDP :: [(KeywordId, DeltaPos)]

Annotations associated with this element.

combineAnns :: Annotation -> Annotation -> Annotation

annNone :: Annotation

type Anns = Map AnnKey Annotation

data AnnKey

For every Located a, use the SrcSpan and constructor name of a as the key, to store the standard annotation. These are used to maintain context in the AP and EP monads

Constructors

AnnKey SrcSpan AnnConName

data KeywordId

We need our own version of keywordid to distinguish between a semi-colon appearing within an AST element and one separating AST elements in a list.

Constructors

G AnnKeywordId

AnnSemiSep

AnnComment DComment

AnnList SrcSpan
In some circumstances we need to annotate a list of statements (e.g. HsDo) and must synthesise a SrcSpan to hang the annotations off. This needs to be preserved so that exactPC can find it, after potential AST edits.

mkAnnKey :: Data a => Located a -> AnnKey

data AnnConName

Constructors

CN

Fields

unConName :: String

annGetConstr :: Data a => a -> AnnConName

data ResTyGADTHook name

Constructors

ResTyGADTHook [LHsTyVarBndr name]

getAnnotationEP :: Data a => Located a -> Anns -> Maybe Annotation

getAndRemoveAnnotationEP :: Data a => Located a -> Anns -> (Maybe Annotation, Anns)

data LayoutFlag

Constructors

LayoutRules

NoLayoutRules