Language.Haskell.GHC.ExactPrint.Utils

Documentation

srcSpanStartLine :: SrcSpan -> Int

srcSpanEndLine :: SrcSpan -> Int

srcSpanStartColumn :: SrcSpan -> Int

srcSpanEndColumn :: SrcSpan -> Int

ss2span :: SrcSpan -> Span

ss2pos :: SrcSpan -> Pos

ss2posEnd :: SrcSpan -> Pos

span2ss :: Span -> SrcSpan

undelta :: Pos -> DeltaPos -> LayoutStartCol -> Pos

Apply the delta to the current position, taking into account the current column offset if advancing to a new line

rdrName2String :: RdrName -> String

isSymbolRdrName :: RdrName -> Bool

deltaFromSrcSpans :: SrcSpan -> SrcSpan -> DeltaPos

Create a delta covering the gap between the end of the first SrcSpan and the start of the second.

ghcCommentText :: Located AnnotationComment -> String

isPointSrcSpan :: SrcSpan -> Bool

ss2deltaP :: Pos -> Pos -> DeltaPos

Convert the start of the second Pos to be an offset from the first. The assumption is the reference starts before the second Pos

isGoodDelta :: DeltaPos -> Bool

isListComp :: HsStmtContext name -> Bool

showGhc :: Outputable a => a -> String

Show a GHC API structure

showAnnData :: Data a => Anns -> Int -> a -> String

For tests

debug :: c -> String -> c

Provide a version of trace the comes at the end of the line, so it can easily be commented out when debugging different things.

debugM :: Monad m => String -> m ()

warn :: c -> String -> c